+++ front-end/src/components/stocks/stockList.css
