+++ front-end/src/features/shops/selectors.ts
