+++ front-end/src/features/shops/shopReducer.ts
