+++ front-end/src/features/shops/shopThunk.ts
