+++ front-end/src/features/shops/types.ts
