+++ front-end/src/features/stocks/actions.ts
