+++ front-end/src/features/stocks/selectors.ts
