+++ front-end/src/features/stocks/stockReducer.ts
