+++ front-end/src/features/stocks/stockThunk.ts
