+++ front-end/src/features/stocks/types.ts
